tools/release-readiness/polarion/requirements.txt
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
distlib==0.3.4
filelock==3.4.1
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gspread==5.1.1
gspread-formatting==1.0.5
httplib2==0.20.2
identify==2.4.4
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.2.3
nodeenv==1.6.0
oauthlib==3.1.1
platformdirs==2.4.0
pre-commit==2.17.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pygsheets==2.0.5
pylero
pyparsing==3.0.7
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
suds-py3==1.4.5.0
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
zipp==3.6.0
jinja2
jinja_markdown
pytz
